requirements.txt
aiogram
pydantic
pydantic-settings
asyncpg
pytz
httpx
PyYAML
redis
